refactor(ff-preview): split playback/mod.rs into focused sub-modules
